requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.28
plotly==5.18.0
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
pillow==10.0.0
requests==2.31.0
streamlit-lottie==0.0.5
streamlit-card==0.0.61
streamlit-extras==0.3.4
streamlit-toggle==0.1.3
streamlit-option-menu==0.3.6
streamlit-folium==0.15.0
altair==5.0.1